Connection started from client. Used threads to keep UI responsive when connected
