+++ src/main/java/ser516/project3/client/service/ClientConnectionServiceImpl.java
@@ -1,33 +1,24 @@
 package ser516.project3.client.service;
 
-import java.io.IOException;
-import java.net.URI;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
-
-import javax.websocket.ContainerProvider;
-import javax.websocket.DeploymentException;
-import javax.websocket.WebSocketContainer;
-
 import org.apache.log4j.Logger;
+
+import ser516.project3.client.helper.ClientConnectionThread;
 
 public class ClientConnectionServiceImpl implements ClientConnectionServiceInterface {
 	final static Logger logger = Logger.getLogger(ClientConnectionServiceImpl.class);
-
-	private static CountDownLatch messageLatch;
+	Thread clientConnectionThread;
 
 	@Override
 	public void createClientConnection(final String ipAddress, final int port, final String endpoint) {
-		messageLatch = new CountDownLatch(1);
-		WebSocketContainer container = ContainerProvider.getWebSocketContainer();
-		String uri = "ws://" + ipAddress + ":" + port + "/" + endpoint;
-		logger.info("Connecting to " + uri);
-		try {
-			container.connectToServer(ClientConnectionEndpoint.class, URI.create(uri));
-			messageLatch.await(100, TimeUnit.SECONDS);
-		} catch (DeploymentException | IOException | InterruptedException e) {
-			logger.error("Exception occurred in createClientConnection method::::" + e.getStackTrace());
-		}
+		ClientConnectionThread threadInstance = new ClientConnectionThread(ipAddress, port, endpoint);
+		clientConnectionThread = new Thread(threadInstance);
+		clientConnectionThread.start();
+	}
+
+	@Override
+	public void stopClientConnection() {
+		clientConnectionThread.interrupt();
+
 	}
 
 }
